+++ boa_engine/src/vm/opcode/delete/mod.rs
@@ -14,7 +14,7 @@ pub(crate) struct DeletePropertyByName;
 impl Operation for DeletePropertyByName {
     const NAME: &'static str = "DeletePropertyByName";
     const INSTRUCTION: &'static str = "INST - DeletePropertyByName";
-    const COST: usize = 3;
+    const COST: u8 = 3;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let index = context.vm.read::<u32>();
@@ -44,7 +44,7 @@ pub(crate) struct DeletePropertyByValue;
 impl Operation for DeletePropertyByValue {
     const NAME: &'static str = "DeletePropertyByValue";
     const INSTRUCTION: &'static str = "INST - DeletePropertyByValue";
-    const COST: usize = 3;
+    const COST: u8 = 3;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let key_value = context.vm.pop();
@@ -72,7 +72,7 @@ pub(crate) struct DeleteName;
 impl Operation for DeleteName {
     const NAME: &'static str = "DeleteName";
     const INSTRUCTION: &'static str = "INST - DeleteName";
-    const COST: usize = 3;
+    const COST: u8 = 3;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let index = context.vm.read::<u32>();
@@ -97,7 +97,7 @@ pub(crate) struct DeleteSuperThrow;
 impl Operation for DeleteSuperThrow {
     const NAME: &'static str = "DeleteSuperThrow";
     const INSTRUCTION: &'static str = "INST - DeleteSuperThrow";
-    const COST: usize = 2;
+    const COST: u8 = 2;
 
     fn execute(_: &mut Context<'_>) -> JsResult<CompletionType> {
         Err(JsNativeError::reference()
