+++ boa_engine/src/vm/opcode/iteration/iterator.rs
@@ -17,7 +17,7 @@ pub(crate) struct IteratorNext;
 impl Operation for IteratorNext {
     const NAME: &'static str = "IteratorNext";
     const INSTRUCTION: &'static str = "INST - IteratorNext";
-    const COST: usize = 6;
+    const COST: u8 = 6;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let mut iterator = context
@@ -46,7 +46,7 @@ pub(crate) struct IteratorFinishAsyncNext;
 impl Operation for IteratorFinishAsyncNext {
     const NAME: &'static str = "IteratorFinishAsyncNext";
     const INSTRUCTION: &'static str = "INST - IteratorFinishAsyncNext";
-    const COST: usize = 5;
+    const COST: u8 = 5;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let mut iterator = context
@@ -85,7 +85,7 @@ pub(crate) struct IteratorResult;
 impl Operation for IteratorResult {
     const NAME: &'static str = "IteratorResult";
     const INSTRUCTION: &'static str = "INST - IteratorResult";
-    const COST: usize = 3;
+    const COST: u8 = 3;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let last_result = context
@@ -114,7 +114,7 @@ pub(crate) struct IteratorValue;
 impl Operation for IteratorValue {
     const NAME: &'static str = "IteratorValue";
     const INSTRUCTION: &'static str = "INST - IteratorValue";
-    const COST: usize = 5;
+    const COST: u8 = 5;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let mut iterator = context
@@ -143,7 +143,7 @@ pub(crate) struct IteratorDone;
 impl Operation for IteratorDone {
     const NAME: &'static str = "IteratorDone";
     const INSTRUCTION: &'static str = "INST - IteratorDone";
-    const COST: usize = 3;
+    const COST: u8 = 3;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let done = context
@@ -170,7 +170,7 @@ pub(crate) struct IteratorReturn;
 impl Operation for IteratorReturn {
     const NAME: &'static str = "IteratorReturn";
     const INSTRUCTION: &'static str = "INST - IteratorReturn";
-    const COST: usize = 8;
+    const COST: u8 = 8;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let record = context
@@ -204,7 +204,7 @@ pub(crate) struct IteratorToArray;
 impl Operation for IteratorToArray {
     const NAME: &'static str = "IteratorToArray";
     const INSTRUCTION: &'static str = "INST - IteratorToArray";
-    const COST: usize = 8;
+    const COST: u8 = 8;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let mut iterator = context
@@ -258,7 +258,7 @@ pub(crate) struct IteratorPop;
 impl Operation for IteratorPop {
     const NAME: &'static str = "IteratorPop";
     const INSTRUCTION: &'static str = "INST - IteratorPop";
-    const COST: usize = 1;
+    const COST: u8 = 1;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         context.vm.frame_mut().iterators.pop();
@@ -276,7 +276,7 @@ pub(crate) struct IteratorStackEmpty;
 impl Operation for IteratorStackEmpty {
     const NAME: &'static str = "IteratorStackEmpty";
     const INSTRUCTION: &'static str = "INST - IteratorStackEmpty";
-    const COST: usize = 1;
+    const COST: u8 = 1;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let is_empty = context.vm.frame().iterators.is_empty();
@@ -295,7 +295,7 @@ pub(crate) struct CreateIteratorResult;
 impl Operation for CreateIteratorResult {
     const NAME: &'static str = "CreateIteratorResult";
     const INSTRUCTION: &'static str = "INST - CreateIteratorResult";
-    const COST: usize = 3;
+    const COST: u8 = 3;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let value = context.vm.pop();
